# special-case `T::Private::Types::SimplePairUnion` in some `T::Utils` code