### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

`T::Utils.unwrap_nilable` and `T::Utils::Nilable.get_underlying_type` turn out to be relatively hot functions in some serialization code in Stripe's codebase.  We can do a better job by adding some special cases for our `SimplePairUnion` specialization.  We can also duplicate a small amount of code so that we don't have to go through `T::Utils.get_type_info` and thereby avoid the allocation of `TypeInfo`.

Benchmarks before:

```
T.unwrap_nilable(Integer): 88.166 ns
get_underlying_type(Integer): 278.329 ns
T.unwrap_nilable(T.nilable(Integer)): 620.187 ns
get_underlying_type(T.nilable(Integer)): 913.333 ns
T.unwrap_nilable(T.any(Float, Integer)): 603.966 ns
get_underlying_type(T.any(Float, Integer)): 843.252 ns
```

After:

```
T.unwrap_nilable(Integer): 92.836 ns
get_underlying_type(Integer): 90.258 ns
T.unwrap_nilable(T.nilable(Integer)): 195.653 ns
get_underlying_type(T.nilable(Integer)): 275.813 ns
T.unwrap_nilable(T.any(Float, Integer)): 156.667 ns
get_underlying_type(T.any(Float, Integer)): 201.998 ns
```

`get_underlying_type` is 3-4x faster; `unwrap_nilable` gets 3-4x faster on `SimplePairUnion`.

### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

See included automated tests.