@froydnj 2023-02-16
OK, this is even a little faster now after the `unwrap_nilable` refactoring.  I'm a little iffy on the logic duplication, but I think it's worth it in the name of performance.